push pointer_chase_benchmarks upstream

+++ hw/block/femu/computation.h
@@ -9,5 +9,5 @@
 #define END_BLOCK_MAGIC 99999
 #define BLOCK_SIZE 4096
 
-int count_bits(char *buf);
-int get_disk_pointer(char *buf);
+uint64_t count_bits(char *buf);
+uint64_t get_disk_pointer(char *buf);
